Use int64 bitboard board_input; update model
Switch board_input arrays from uint8 to int64 and encode per-square legal-to bitboards (add _as_int64). Update board_to_input to return legal-to bitboards + last_from. Adapt dataset, self_play, and self_play_game to the new dtype. Revise ChessNet: remove meta_positions/meta_pos_emb, replace small embeddings for legal/last-from with a legal_to_mlp and legal_from linear, add bit_positions buffer and unpack bitboards in forward. Integrates bitboard-based legal move encoding into the model input pipeline.

+++ data_generation.py
@@ -197,7 +197,7 @@ def position_label(
 
     total = sum(pair_scores.values()) or 1.0
     return {
-        "board_input": np.array(board_to_input(board), dtype=np.uint8),
+        "board_input": np.array(board_to_input(board), dtype=np.int64),
         "legal_pairs": np.array(
             list(legal_moves_by_square_pair(board, legal_moves=legal_moves).keys()),
             dtype=np.uint8,
